+++ agents/cap_group_agents/Huawei_Cloud_API_group/Huawei_Cloud_API_manager.py
@@ -1,5 +1,5 @@
 from agency_swarm import Agent
-from agents.tools.read_json_file.read_json_file import ReadJsonFile
+from agents.tools.read_json_file.ReadJsonFile import ReadJsonFile
 
 _name = "Huawei_Cloud_API_mananger"
