+++ agents/cap_group_agents/IAM_service_group/IAM_service_manager.py
@@ -1,5 +1,5 @@
 from agency_swarm import Agent
-from agents.tools.read_json_file.read_json_file import ReadJsonFile
+from agents.tools.read_json_file.ReadJsonFile import ReadJsonFile
 
 _name = "IAM_service_mananger"
 
@@ -64,7 +64,7 @@ _instruction = f"""
 """
 
 
-_tools = []
+_tools = [ReadJsonFile]
 
 _file_folder = ""
