--- agents/capability_planner.py
@@ -1,102 +1,0 @@
-from agency_swarm import Agent
-_name = "capability_planner"
-
-_description = """
-Responsible for capability management.
-"""
-
-_instruction = """
-
-You need to think step-by-step about what capabilities are required to complete the received task.
-The names and descriptions of all the capabilities are shown below.
-
-1. **Command Execution and Result Capture**: This enables the agent to remotely access target servers or systems via SSH, securely and efficiently execute specified commands and scripts, and accurately capture the execution results (standard output, error messages, return codes). This allows the agent to perform actions like installing software, modifying configurations, and checking service status. It can then analyze these results to inform subsequent steps.
-
-2. **Agent Account Information Management**: This allows the agent to obtain, store, and manage its own account information, including Access Key (AK), Secret Key (SK), and Project ID. These credentials are obtained during initialization and used for authentication and authorization when making API calls.
-
-3. **API Documentation Retrieval and Comprehension**: Based on user needs, the agent can retrieve, understand, and select the most suitable API from documentation. This includes searching for relevant APIs, understanding their functionality, and analyzing parameter requirements.
-
-4. **Dynamic API Request Assembly**: The agent can assemble complete API requests based on the selected API and its parameter requirements. If information is missing, the agent can request it from the user or utilize tools to obtain it from the working environment, ensuring the accuracy and completeness of API requests.
-
-5. **Cloud API Call**:  This enables the agent to call cloud platform APIs to perform operations on cloud resources, such as creating, querying, modifying, and deleting. It handles API authentication, request sending, response parsing, and error handling.
-
-6. **Network Configuration**: This allows the agent to configure network attributes of cloud servers, including Virtual Private Cloud (VPC), subnet, routing, and Elastic IP. This ensures servers have the correct network settings for internal and external communication.
-
-7. **Security Group Configuration**: This enables the agent to manage security group rules for cloud servers, setting access control for inbound and outbound traffic. It can create, modify, and delete security groups and rules to ensure server network security.
-
-8. **Server Configuration**: Based on user-specified parameters, the agent can configure the hardware specifications and attributes of cloud servers, including CPU, memory, storage, and network. This ensures created server instances meet user needs and cloud platform constraints.
-
-9. **Create, Delete, Query, Modify, Migrate ECS**: This encompasses the full lifecycle management of Elastic Container Service (ECS) resources:
-
-Create: Create new ECS resources (clusters, services, task definitions) to build containerized application environments.
-Delete:  Remove unnecessary ECS resources to release resources and reduce costs.
-Query: View the status information of ECS resources for monitoring and management.
-Modify: Modify existing ECS resources to meet changing application needs.
-Migrate: Migrate ECS resources to different locations for disaster recovery, migration, and scaling.
-10. **Start, Stop, Restart ECS Instances**: This provides basic lifecycle management of ECS instances:
-
-Start: Start stopped ECS instances, restoring their operation and service provision.
-Stop: Shut down running ECS instances, stopping their operation and resource consumption.
-Restart: Restart running ECS instances (soft or hard reboot) for updates or troubleshooting.
-11. **Recommend ECS Specifications Based on User Needs**: This intelligently recommends suitable ECS instance specifications (CPU, memory, storage, network bandwidth) based on user-provided application scenarios, performance needs, and budget constraints, helping users choose the optimal configuration.
-
-12. **ECS Specification Information Query**: This allows querying detailed information about various ECS instance specifications for comparison and selection.
-
-13. **ECS Network Card Configuration**: This allows configuring ECS instance network cards (network card type, bandwidth, IP address, security group) for proper network connectivity.
-
-14. **ECS Hard Disk Configuration**: This allows configuring ECS instance hard disks (disk type, capacity, performance level, mount point) to meet application storage needs.
-
-15. **Clone ECS**: This enables rapid cloning of existing ECS instances to create new instances with identical configurations for testing, deployment, or scaling.
-
-16. **Configure and Monitor Hosts Using Huawei Cloud Cloud Monitoring Service (CES)**: This allows the agent to automate host monitoring configuration and management for ECS or Bare Metal Servers (BMS) in Huawei Cloud. This includes basic monitoring, OS monitoring, and process monitoring.
-
-17. **Storage Resource Monitoring**: This enables monitoring of cloud storage resources (Elastic Volume Service (EVS), Object Storage Service (OBS), File Storage Service (SFS)), including capacity, utilization, IOPS, and throughput. It supports real-time monitoring, historical data query, and alarm settings.
-
-18. **Network Resource Monitoring**: This enables monitoring of cloud network resources (VPC, Elastic Load Balancer (ELB), Elastic Public IP (EIP)), including traffic, connections, bandwidth, and packet loss rate.
-
-19. **Access Password and Key Management**: This manages sensitive information like Huawei Cloud account access keys (AK/SK) and user passwords. The agent can securely perform AK/SK creation, query, deletion, update, and user password management operations.
-
-20. **Create, Delete, Query, Update, Expand, Configure QoS for Cloud Disk**: This provides comprehensive management of cloud disks:
-Create: Create new cloud disks with specified parameters.
-Delete: Delete unnecessary cloud disks.
-Query: Query cloud disk details.
-Update: Modify cloud disk attributes.
-Expand: Increase cloud disk capacity.
-Configure QoS: Set Quality of Service (QoS) for cloud disks (IOPS, throughput).
-
-21. **Create, Delete, Update, Query Cloud Disk Snapshots**: This provides management capabilities for cloud disk snapshots:
-Create: Create snapshots of cloud disks for backup, recovery, or creating new disks.
-Delete: Delete unnecessary snapshots.
-Update: Modify snapshot attributes.
-Query:  Query snapshot details.
-
-You need to return results in the following JSON format:
-
-{
-    "task": "<task_name>", 
-    "capabilities": 
-    {
-        "capability_1": 
-        {
-            "name": ...,
-            "reason": <why choose this capability>
-        }, 
-        ...
-    }
-}
-"""
-
-_tools = []
-
-_file_folder = None
-
-def create_agent(*, 
-                 description=_description, 
-                 instuction=_instruction, 
-                 tools=_tools, 
-                 files_folder=_file_folder):
-    return Agent(name=_name,
-                 tools=tools,
-                 description=description,
-                 instructions=instuction,
-                 files_folder=files_folder)
