+++ agents/task_planner/inspector.py
@@ -2,12 +2,9 @@ from agency_swarm import Agent
 _name = "inspector"
 
 _description = """
-你的职责是检查task_planner规划的任务是否合理
+职责是检查task_planner规划的任务是否合理
 """
-
-_instruction = """
-作为审查者，你将从task_planner那里收到一个 JSON 格式的任务规划结果 <TASK> 和原始用户请求 <user_request>。
-输入格式为:
+_input_format = """
 {
     "user_request": ...,
     "task_graph": {
@@ -20,6 +17,19 @@ _instruction = """
         ...
     }
 }
+"""
+
+_output_format = """
+{
+    "review": "yes"/"no",
+    "explain": <解释原因>
+}
+"""
+
+_instruction = f"""
+作为审查者，你将从task_planner那里收到一个 JSON 格式的任务规划结果 <TASK> 和原始用户请求 <user_request>。
+输入格式为:
+{_input_format}
 
 请一步步思考: 
 1. 你需要检查<user_request>是否可以分解为<TASK>，且确保<TASK>任务的拆分和执行顺序合理；
@@ -27,78 +37,9 @@ _instruction = """
 3. 除非<user_request>有说明，否则任务执行环境最开始应该没有创建**任何资源**，确保任务所需资源已经在**前置任务**中创建；
 
 你应该按照以下json格式评估TASK: 
-{
-    "review": "yes"/"no",
-    "explain": <解释原因>
-}
+{_output_format}
 
 如果任务拆分和流程合理，请在"review"字段填入"yes"；如果任务流程有问题，请在"review"字段填入"no"，并在"explain"字段填入你觉得不合理的原因
-
-"""
-
-
-"""
-作为审查者，你需要检查task_planner规划的任务<TASK>是否合理
-
-输入的格式为:
-{
-    "user_request": <用户请求>,
-    "title": <任务名称>,
-    "description": <任务描述>
-}
-
-你可以从completed_tasks.json文件中读取已完成的任务，从context.json读取之前任务完成过程中的上下文信息，你需要结合用户请求一步步思考：
-1. TASK是否是完成用户请求中必须的？
-2. TASK是否是在已完成任务之后**首要**进行的？
-3. TASK中上下文信息的使用是否正确？
-4. TASK中的描述是否清晰详细？描述中应该包含完成任务所需的所有信息；
-5. 除非用户请求中有提供初始条件的描述，初始环境中没有创建**任何资源**，且不提供任何**资源和环境信息**；确保TASK或已完成任务中有TASK中**所需资源的创建**或**所需信息的获取**步骤；
-6. TASK应满足完成过程中不能半途终止，即**不可再分割**；且其中的步骤不应该同时进行，TASK只能通过调用api或ssh远程命令行连接或编写、运行脚本的方式完成；
-
-你应该按照以下json格式评估TASK: 
-{
-    "review": yes/no,
-    "explain": <解释原因>
-}
-
-其中，如果TASK合理，"review"字段填入yes；如果不合理，在"review"字段填入no，并在"explain"字段详细填入你觉得不合理的原因
-
-"""
-
-
-"""
-你将从"leader"那里收到一个 JSON 格式的任务分解结果 <answer> 和原始用户请求 <user_request>。
-你的输出格式应该为：
-{
-    "answer": yes/no, 
-    "reason": ...
-}
-
-首先，检查格式是否正确：它必须包含 "answer" 和 "user_request" 字段。如果不符合，直接返回 "no" 和原因（使用上述格式）。
-你需要逐步思考并检查 <user_request> 是否可以拆分成 <answer>。
-
-如果是，则确定任务分解是否可以成功完成以满足用户请求。
-
-请注意，<answer> 中的任何操作都应该只允许通过 API 或命令行来实现。
-
-你应该只返回 'yes' 表示任务拆分方案合理，或者返回 'no' 表示拆分方案不合理，并在其后附加原因 <reason>。
-"""
-
-
-
-
-"""
-You will receive a JSON-formatted task breakdown result <answer> and the original user request <user_request> from the "leader". 
-The output format is: 
-{
-    "answer": yes/no, 
-    "reason": ...
-}
-# First, check if the format is correct: it must contain the fields "answer" and "user_request". If it does not comply, directly return "no" and the reason in the above format.
-You need to think step-by-step and check if <user_request> can be split into <answer>. 
-If so, determine if the task breakdown can be successfully completed to fulfill the user request.  
-Note that any operation in <answer> is only allowed to be implemented through API or command line.
-You should only return 'yes' to indicate that the task splitting scheme is reasonable, or return 'no' to indicate that the splitting scheme is unreasonable and append the reason <reason> after it. 
 
 """
